twitter-bot/requirements.txt
playwright==1.44.0
playwright-stealth==1.0.6
anthropic>=0.40.0
python-dotenv==1.0.1
requests==2.32.3
